fix: Implement real verification in TestAssertions and refactor test infrastructure
- Update DrumTrack.notesCount to exclude deleted notes, mirroring safeNotes semantics
- Replace placeholder TestAssertions.assertDeleted/assertNotDeleted with real verification using model.isDeleted property
- Add ModelContext parameter to assertion helpers for proper database verification
- Update all test call sites to pass context parameter for accurate deletion checking
- Split ServerSongModelTests into separate ServerChartModelTests and ServerSongModelTests to fix SwiftLint type_body_length violations
- Remove unused createSong helper from ContentViewTests to eliminate dead code
- Convert all tests to use TestContainer.shared.context and TestSetup.withTestSetup for proper test isolation
- Ensure duplicate-removal regressions are caught with meaningful assertion failures instead of false positives
